Moved utility functions out of `bazel` and into `util`
